requirements.txt
blinker==1.3
certifi==2017.7.27.1
chardet==2.3.0
cloud-init==0.7.9
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
future==0.16.0
gpxpy==1.1.2
idna==2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
nltk==3.2.5
oauthlib==1.0.3
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
python-telegram-bot==8.1.1
PyYAML==3.11
requests==2.9.1
six==1.10.0
ssh-import-id==5.5
tracery==0.1.1
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
xmltodict==0.11.0
